move cors functions to helper in router
